requirements.dev.txt
ipython
flake8
flake8-isort
flake8-docstrings
flake8-builtins
flake8-quotes
flake8-bugbear
flake8-comprehensions
flake8-eradicate
flake8-pep3101
flake8-print
flake8-super
flake8-django
pytest==6.2.4
pytest-cov==2.11.1
pytest-django==4.2.0
pytest-factoryboy==2.1.0
autopep8==1.5.7
black==21.5b1